Fuel Counter - C
A gas station wants to know which is the most consumed fuel. The user enters the codes of each fuel, which are 1 = alcohol, 2 = gasoline, 3 = diesel. The algorithm will sum how many times each fuel was typed, then prints a report at the end. The program will finish when the user enter the number 4.